+++ packages/components/pie-cookie-banner/src/defs.ts
@@ -1,6 +1,8 @@
 import { type TemplateResult } from 'lit';
 import { type ComponentDefaultProps } from '@justeattakeaway/pie-webc-core';
 
+import defaultLocale from '../locales/en-gb.json' assert { type: 'json' };
+
 export interface CookieBannerLocale {
     banner: {
         title: string;
@@ -141,3 +143,11 @@ export interface CustomTagEnhancers {
 }
 
 export type DefaultProps = ComponentDefaultProps<CookieBannerProps>;
+
+export const defaultProps: DefaultProps = {
+    hasPrimaryActionsOnly: false,
+    defaultPreferences: {},
+    locale: defaultLocale,
+    cookieStatementLink: '',
+    cookieTechnologiesLink: '',
+};