+++ apps/examples/wc-react17/CHANGELOG.md
@@ -1,5 +1,12 @@
 # wc-react17
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.5.18
+
 ## 0.5.2
 
 ### Patch Changes